Updated functionality to fetch data from "/rest/players" endpoint
Updated a function to send a GET request to the "/rest/players" URL and append the resulting data to the table specified in point 7 of the requirements. Upon restarting the server, the table now displays four rows: the first row contains column names, and the subsequent rows contain data fetched from the server.