Implemented edit functionality for player records
Added functionality to edit player records when clicking the "Edit" button. This includes changing the icon and enabling editing for the following fields: Name, Title, Race, Profession, and Banned. The fields are pre-populated with the current values from the player record.

Implemented sending changes to the server upon clicking the "Save" button. Utilized a POST request to the "/rest/players/{id}" endpoint with the updated field values in JSON format using JSON.stringify({}). After making changes, updated the data on the current page.